Moved Libarary Source out of Project and Added drawSprites function
Originally I had the library source in the actual project folder, I moved it out to the main repo folder.
Also added a drawSprites function because calling .draw() on every sprite is fine when there's only 3 sprites but it would get pretty tedious pretty quick as we start to scale up.