Sprite Text Implementation
This took a bit longer than it needed to because I implemented this with the plan I have to add instancing to sprites. All (I've only tested 2 at once) text is drawn with a single draw call. I believe nearly the exact same method can be applied to sprites, and then we'll have 3 draw calls total: Sprites, DebugDraw, and SpriteText