Uncommented testing.cpp
So I'm actually not using this file for anything right now, these are old GFX main functions that arent actually called by anything. However, it was only all commented when I tried my first implementation of instanced sprites and this caused everything to break. So we're not using that instanced sprite implementation anymore, so I uncommented all this