Sprites' color can be tinted
We dont have color for debug draw or spriteText right now. I think im going to need to write a new shader for debugDraw, and since the text is black it's just not gonna tint. I'm probably gonna generate another font sprite sheet with white text so it can be tinted.