Atlas now assumes 1 frame
I wanted to test the atlas more so I added the background, but I forgot to add the _1 to the end of it. So when I ran the game it crashed. Which seemed pretty bad. So now, if there's not a _NumFrames at the end of a filename, it'll assume the file is only 1 frame and the game won't crash.