Lots of stuff
Cleaned up the camera because I thought it was  necessary for figuring out world to camera.
Added some constructors to vector 3 to make it more convient for myself to use
Tried for a while to get the mouse vector thing working. Got closer for sure, but still not unfortunately.
Worked on mouse vector until the art was posted, then worked on implementing the art for a bit.
It's almost implemented but not quite fully yet, I'm too tired to continue.