Functional and improved headerizer
Wont stackdump if you try to put in remarks anymore
Now adds a ~EndHeader~ tag to end of header so you can modify the headers of stuff that already has headers. I'm gonna have to go through and manually remove all the old generated headers because without the tag there's no way for it to distinguish between its programmatically generated header and actual code. So. That tag exists now.
Added word wrapping to the brief and remarks section so that it doesnt go beyond column 78