Tilemap improvements
I'm still at school for 90 mins but i think im gonna just chill. Not feeling great. This commit allows you to select a texture through the editor. Tried to get saving recent textures fixed but its not workiing and I just cant be bothered to fix it right now. I might work more tonight and if not ill be back at it tomorrow.