Fixed issue where mouseVector and Playerstate couldn't be added.
Playerstate was saving a tilemapcollision* but never using it. I commented out the ptr and it worked.
Mousevector was dynamically allocating a sound emitter in it ctor?! Don't do this. Also, if you test in someone else's script, and you're not adding to the script, just testing in it... Remove the tests when you're done! One file, one purpose.  Really if you're testing functionality you should be writing your own functions or classes to store that functionality anyway. I left it commented out so you have it as a reference though.
Physics (rigidbody and shape) still cannot be added at run time. I thought they could, but they can't. It looks like you successfully added them, but it crashes the game the next time you try to add a component. Please remove mutual dependency.