Entity Loading and tons of other stuff
FOR JOSH/MITCH - I think, I'm not 100% sure, but I think the only thing that doesn't work at this point is remove component. Also if you try to add a body or AAB it'll silently not work I think. It might give an error message. The biggest issues right now are with reactive's dual role as an Editor necessity and a useful component. Maybe we just want to sep into editor reactive and reactive? They can inherit from the same reactive base. I think the true reactive stuff i added is rubbish and might not work but i need to stop work for the night at some point.

Other stuff:
-Added true reactive designator to reactive to determine whether or not the editor automatically added the reactive component. Don't think this fully works right now. I'll look more at later. Shouldn't affect editor at all.
-Fixed reactive. The logic in reactive was such that negative scales would cause it to always return false for all mouse states
-Moved gamelogic to the final system, just for safety.
-Fixed sprite's serialize read, was trying to save a string as an AtlasTexture*
-Moved some editor entity stuff to factory so its more universally accessible.
-Removed camera serialize. I dont think we ever actually want to save its state.