Can now add physics to entities with editor
Although the tilemap bug still persists! I tried to also add that tilemap functionality where it gets a rigid body if it didn't have one but that still breaks. Oh well, not a big deal.
Also, when I first added body and shape, the obj wasn't affected by grav. Default ctor for body was setting mass and restitution = 0, so I changed that

~20min