More tweakables and a quality of life improvement
Transform is now fully tweakable, not just mouse controls
A sprite's color is now technically tweakable, but the second you change it it'll immediately be overwritten by reactive green, which I'll fix when I fix reactive.
Added a quality of life change where nothing can be selected or changed via mouse while your mouse is over the editor window. The days of adding a random tile because It happened to be under the save level button are over.