Visual Layers
First of all, shoutout to stl for making this really easy to implement.
The front layer is 0 and the back layer is NUMLAYERS, which is currently set to 5. Theoretically, this supports an infinite amount of layers.
The default for sprites is to be set to NUMLAYERS - 1, the back layer.
Tested this with the hud component, setting it to layer 1, the second layer from the front. Figured the frontmost layer would probably be for debug stuff.
Behavior of overlapping objects in same layer is whichever object updated last draws on top. But with infinite layer support this should never be an issue.

~45 min